added error handling

+++ src/main/java/foodraport/foodraport/api/controller/UserController.java
@@ -2,8 +2,10 @@ package foodraport.foodraport.api.controller;
 
 import foodraport.foodraport.api.model.User;
 import foodraport.foodraport.service.UserService;
+import org.springframework.http.HttpStatus;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -21,18 +23,23 @@ public class UserController {
         this.userService = userService;
     }
 
-    @GetMapping("/user/{id}")
-    public User getUser(@RequestParam Integer id) {
+    @GetMapping("/user/")
+    public ResponseEntity<?> getUser(@RequestParam Integer id) {
         User user = userService.getUser(id);
         if (user != null) {
-            return user;
+            return ResponseEntity.status(HttpStatus.OK).body(user);
         }
-        return null;
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User id " + id + " not found.");
     }
 
-    @PutMapping("/user/{id}")
-    public User updateUser(@RequestParam Integer id, @RequestParam String foods) {
+    @PutMapping("/user/")
+    public ResponseEntity<?> updateUser(@RequestParam Integer id, @RequestParam String foods) {
         User user = userService.getUser(id);
+
+        if (user == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User id " + id + " not found.");
+        }
+        
         String existingFoods = userService.getUser(id).getFoods();
         String newFoods = foods;
         if (user != null) {
@@ -45,17 +52,30 @@ public class UserController {
             }
         
         user.setFoods(foods);
-        return user;
+        return ResponseEntity.status(HttpStatus.OK).body(user);
     }
 
-    @PostMapping("/user/{id}")
-    public User addUser(@RequestParam Integer id, @RequestParam String name, @RequestParam String email, @RequestParam String foods) {
-        User user = userService.addUser(id, name, email, foods);
-        return user;
+    @PostMapping("/user/")
+    public ResponseEntity<?> addUser(@RequestParam Integer id, @RequestParam String name, @RequestParam String email, @RequestParam String foods) {
+        User user = userService.getUser(id);
+        
+        if (user != null){
+        return ResponseEntity.status(HttpStatus.CONFLICT).body("User id " + id + " already in use. Please use another id.");
     }
 
-    @DeleteMapping("/user/{id}")
-    public void deleteUser(@RequestParam Integer id) {
+        userService.addUser(id, name, email, foods);
+
+        return ResponseEntity.status(HttpStatus.CREATED).body("User with id " + id + " created.");
+    }
+
+    @DeleteMapping("/user/")
+    public ResponseEntity<?> deleteUser(@RequestParam Integer id) {
+        User user = userService.getUser(id);
+        
+        if (user == null){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User with id " + id + " not found.");
+        }
         userService.deleteUser(id);
+        return ResponseEntity.status(HttpStatus.OK).body("User with id " + id + " deleted.");
     }
 }
